+++ racetrack_commons/racetrack_commons/api/asgi/asgi_server.py
@@ -1,7 +1,8 @@
 import contextlib
+import signal
 import threading
 import time
-from typing import Union
+from typing import Union, Callable, Optional
 import logging
 
 import uvicorn
@@ -9,6 +10,7 @@ from uvicorn.config import LOGGING_CONFIG
 from uvicorn.logging import DefaultFormatter, AccessFormatter
 from starlette.types import ASGIApp
 
+from racetrack_client.log.exception import log_exception
 from racetrack_client.log.logs import get_logger
 from racetrack_commons.api.debug import is_deployment_local
 
@@ -26,7 +28,7 @@ HIDDEN_ACCESS_LOGS = {
     'GET /metrics/ 200',
 }
 
-UVICORN_DEBUG_LOGS = False
+UVICORN_ERROR_LOGS = True
 
 
 def serve_asgi_app(
@@ -34,12 +36,36 @@ def serve_asgi_app(
     http_port: int,
     http_addr: str = '0.0.0.0',
     access_log: bool = False,
+    on_shutdown: Optional[Callable[[], None]] = None
 ):
     use_reloader = is_deployment_local() and isinstance(app, str)
     mode_info = ' in RELOAD mode' if use_reloader else ''
     logger.info(f'Running ASGI server on http://{http_addr}:{http_port}{mode_info}')
     _setup_uvicorn_logs(access_log)
-    uvicorn.run(app=app, host=http_addr, port=http_port, log_level="debug", reload=use_reloader)
+
+    config = uvicorn.Config(
+        app=app,
+        host=http_addr,
+        port=http_port,
+        log_level="debug",
+        reload=use_reloader,
+        timeout_graceful_shutdown=3,
+    )
+    server = ManagableServer(config)
+
+    def shutdown_signal_handler(sig, frame):
+        logger.info(f'received signal {sig}, shutting down...')
+        if on_shutdown is not None:
+            try:
+                on_shutdown()
+            except BaseException as e:
+                log_exception(e)
+        server.should_exit = True
+
+    signal.signal(signal.SIGTERM, shutdown_signal_handler)
+    signal.signal(signal.SIGINT, shutdown_signal_handler)
+
+    server.run()
 
 
 def serve_asgi_in_background(
@@ -50,7 +76,13 @@ def serve_asgi_in_background(
 ) -> contextlib.AbstractContextManager:
     logger.info(f'Running ASGI server in background on http://{http_addr}:{http_port}')
     _setup_uvicorn_logs(access_log)
-    config = uvicorn.Config(app=app, host=http_addr, port=http_port, log_level="debug")
+    config = uvicorn.Config(
+        app=app,
+        host=http_addr,
+        port=http_port,
+        log_level="debug",
+        timeout_graceful_shutdown=3,
+    )
     return BackgroundServer(config=config).run_in_thread()
 
 
@@ -73,8 +105,8 @@ def _setup_uvicorn_logs(access_log: bool):
         },
     }
 
-    if not UVICORN_DEBUG_LOGS:
-        LOGGING_CONFIG["loggers"]["uvicorn"]["propagate"] = False
+    LOGGING_CONFIG["loggers"]["uvicorn"]["propagate"] = False
+    if not UVICORN_ERROR_LOGS:
         LOGGING_CONFIG["loggers"]["uvicorn.error"]["level"] = 'INFO'
         LOGGING_CONFIG["loggers"]["uvicorn.error"]["propagate"] = False
 
@@ -140,3 +172,8 @@ class BackgroundServer(uvicorn.Server):
         finally:
             self.should_exit = True
             thread.join()
+
+
+class ManagableServer(uvicorn.Server):
+    def install_signal_handlers(self):
+        pass